This commit contains a restructured codebase as well as code related to new features being implemented this is not in a fully working order and would not be committed this early but due to the restructured file layout this code has been pushed.
Core changes to how we record Entity Hurt Events and the start of being able to play those back this will allow for PVP to be recorded and played back. Currently this does not work and is being worked on.

Types have been updated to store the new event data that required.